extract token validation
